shuffle things around slightly
