+++ src/shared/Component.js
@@ -1,64 +1,7 @@
-function noop() {}
+import { Base } from './Base.js';
+import { assign, callAll, noop } from './utils.js';
 
-function assign(tar, src) {
-	for (var k in src) tar[k] = src[k];
-	return tar;
-}
-
-function _differsImmutable(a, b) {
-	return a != a ? b == b : a !== b;
-}
-
-function callAll(fns) {
-	while (fns && fns.length) fns.shift()();
-}
-
-function blankObject() {
-	return Object.create(null);
-}
-
-class Base {
-	constructor() {
-		this._handlers = blankObject();
-	}
-
-	fire(eventName, data) {
-		const handlers = eventName in this._handlers && this._handlers[eventName].slice();
-		if (!handlers) return;
-
-		for (let i = 0; i < handlers.length; i += 1) {
-			const handler = handlers[i];
-
-			if (!handler.__calling) {
-				handler.__calling = true;
-				handler.call(this, data);
-				handler.__calling = false;
-			}
-		}
-	}
-
-	get() {
-		return this._state;
-	}
-
-	on(eventName, handler) {
-		const handlers = this._handlers[eventName] || (this._handlers[eventName] = []);
-		handlers.push(handler);
-
-		return {
-			cancel: function() {
-				const index = handlers.indexOf(handler);
-				if (~index) handlers.splice(index, 1);
-			}
-		};
-	}
-
-	_differs(a, b) {
-		return _differsImmutable(a, b) || ((a && typeof a === 'object') || typeof a === 'function');
-	}
-}
-
-class Component extends Base {
+export class Component extends Base {
 	constructor(options) {
 		super();
 		this._init(options);
@@ -135,39 +78,30 @@ class Component extends Base {
 	}
 }
 
-/* generated by Svelte vX.Y.Z */
-
-function create_main_fragment(component, state) {
-
-	return {
-		c: noop,
-
-		m: noop,
-
-		p: noop,
-
-		u: noop,
-
-		d: noop
-	};
-}
-
-class SvelteComponent extends Component {
+export class ComponentDev extends Component {
 	constructor(options) {
-		super(options);
-		this._state = assign({}, options.data);
-
-		this._fragment = create_main_fragment(this, this._state);
-
-		if (options.target) {
-			this._fragment.c();
-			this._mount(options.target, options.anchor);
+		if (!options || (!options.target && !options.root)) {
+			throw new Error(`'target' is a required option`);
 		}
+
+		super(options);
 	}
 
-	foo() {
-		console.log('foo');
+	destroy(detach) {
+		super.destroy(detach);
+		this.destroy = () => {
+			console.warn('Component was already destroyed');
+		};
 	}
+
+	set(newState) {
+		if (typeof newState !== 'object') {
+			throw new Error(`${this._debugName}.set was called without an object of data key-values to update.`);
+		}
+
+		this._checkReadOnly(newState);
+		super.set(newState);
+	}
+
+	_checkReadOnly() {}
 }
-
-export default SvelteComponent;
