+++ src/shared/SvelteElement.js
@@ -1,0 +1,5 @@
+class SvelteElement extends HTMLElement {
+	constructor(options) {
+
+	}
+}
